Simplify: shared feature list, header-box tokens, leaner hot paths
Reuse: the remote-compute explainer's data, rows, and styles become one
shared feature-list module (Settings and the onboarding wizard render
the same story); peerRowStyles is promoted out of the settings grab-bag
to styles/peer-rows.ts; the peer pill/paired-ago logic moves to
util/peer-display.ts; the M:SS formatter joins its relative-time
siblings as formatMinSec; dead listener-badge CSS and unused banner-*
class names go.

Altitude: the header's logo/kebab footprints become shared tokens
(--esphome-header-logo-box/-action-box) consumed by the header and the
stack bars — which also fixes a 6px chevron/icon misalignment above the
870px breakpoint where the logo is 44px; the discovery banner's stacks
variant moves beside its base rules in device-grid-styles; the
viewport-height calcs gain the codified vh fallback; the stacks
attributes toggle unconditionally so the changed-key list (which had
already missed tour state) can't silently go stale.

Efficiency: the device grid is built lazily via a content thunk so a
collapsed builder section never constructs the O(devices) template it
would discard; both panel tickers suspend while collapsed (the pairing
countdown via a new setTickSuspended that keeps the baseline anchored);
tour session state is read once per render cycle instead of per getter
call.

Simplification: one _onSwapStacks handler replaces the two identical
toggles, the YAML and cards/table branches share one _renderStacked
scaffold, and the session store keeps a bare string instead of a JSON
envelope (dropping its shape validation and own-key guard).
