+++ src/util/dashboard-stacks-session.ts
@@ -11,17 +11,10 @@ export const STORAGE_KEY = "esphome-dashboard-stacks";
 
 export type DashboardStack = "remote" | "builder";
 
-/** Read the session's expanded-stack choice; null = no choice yet. */
+/** Read the session's expanded-stack choice; null = no (valid) choice yet. */
 export function loadExpandedStack(): DashboardStack | null {
   try {
-    const raw = sessionStorage.getItem(STORAGE_KEY);
-    if (raw === null) return null;
-    const parsed: unknown = JSON.parse(raw);
-    if (parsed === null || typeof parsed !== "object" || Array.isArray(parsed))
-      return null;
-    // Own-key check so a polluted Object.prototype can't supply the value.
-    if (!Object.prototype.hasOwnProperty.call(parsed, "expanded")) return null;
-    const value = (parsed as Record<string, unknown>).expanded;
+    const value = sessionStorage.getItem(STORAGE_KEY);
     return value === "remote" || value === "builder" ? value : null;
   } catch {
     return null;
@@ -31,7 +24,7 @@ export function loadExpandedStack(): DashboardStack | null {
 /** Persist the expanded-stack choice; drops the write if storage is unavailable. */
 export function saveExpandedStack(stack: DashboardStack): void {
   try {
-    sessionStorage.setItem(STORAGE_KEY, JSON.stringify({ expanded: stack }));
+    sessionStorage.setItem(STORAGE_KEY, stack);
   } catch {
     // Drop the write; the swap still works for this render.
   }
